+++ Banks/Transactions/TransactionReplenishment.cs
@@ -5,7 +5,7 @@ namespace Banks
 {
     public class TransactionReplenishment : Transaction
     {
-        public TransactionReplenishment(Account sender, Account recipient, float transactionAmount, int id)
+        public TransactionReplenishment(Account sender, Account recipient, double transactionAmount, int id)
             : base(id, sender, recipient, transactionAmount)
         {
             recipient.IncreaseMoney(transactionAmount);
